chore: organize code
