+++ prowler/lib/outputs/models.py
@@ -1,132 +1,9 @@
-import sys
 from typing import List, Optional
 
 from pydantic import BaseModel
 
-from prowler.lib.logger import logger
 
-
-def get_check_compliance(finding, provider_type, output_options) -> dict:
-    """get_check_compliance returns a map with the compliance framework as key and the requirements where the finding's check is present.
-
-        Example:
-
-    {
-        "CIS-1.4": ["2.1.3"],
-        "CIS-1.5": ["2.1.3"],
-    }
-    """
-    try:
-        check_compliance = {}
-        # We have to retrieve all the check's compliance requirements
-        if finding.check_metadata.CheckID in output_options.bulk_checks_metadata:
-            for compliance in output_options.bulk_checks_metadata[
-                finding.check_metadata.CheckID
-            ].Compliance:
-                compliance_fw = compliance.Framework
-                if compliance.Version:
-                    compliance_fw = f"{compliance_fw}-{compliance.Version}"
-                if compliance.Provider == provider_type.upper():
-                    if compliance_fw not in check_compliance:
-                        check_compliance[compliance_fw] = []
-                    for requirement in compliance.Requirements:
-                        check_compliance[compliance_fw].append(requirement.Id)
-        return check_compliance
-    except Exception as error:
-        logger.critical(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}] -- {error}"
-        )
-        sys.exit(1)
-
-
-def unroll_list(listed_items: list):
-    unrolled_items = ""
-    separator = "|"
-    if listed_items:
-        for item in listed_items:
-            if not unrolled_items:
-                unrolled_items = f"{item}"
-            else:
-                unrolled_items = f"{unrolled_items} {separator} {item}"
-
-    return unrolled_items
-
-
-def unroll_tags(tags: list):
-    unrolled_items = ""
-    separator = "|"
-    if tags and tags != [{}] and tags != [None]:
-        for item in tags:
-            # Check if there are tags in list
-            if isinstance(item, dict):
-                for key, value in item.items():
-                    if not unrolled_items:
-                        # Check the pattern of tags (Key:Value or Key:key/Value:value)
-                        if "Key" != key and "Value" != key:
-                            unrolled_items = f"{key}={value}"
-                        else:
-                            if "Key" == key:
-                                unrolled_items = f"{value}="
-                            else:
-                                unrolled_items = f"{value}"
-                    else:
-                        if "Key" != key and "Value" != key:
-                            unrolled_items = (
-                                f"{unrolled_items} {separator} {key}={value}"
-                            )
-                        else:
-                            if "Key" == key:
-                                unrolled_items = (
-                                    f"{unrolled_items} {separator} {value}="
-                                )
-                            else:
-                                unrolled_items = f"{unrolled_items}{value}"
-            elif not unrolled_items:
-                unrolled_items = f"{item}"
-            else:
-                unrolled_items = f"{unrolled_items} {separator} {item}"
-
-    return unrolled_items
-
-
-def unroll_dict(dict: dict):
-    unrolled_items = ""
-    separator = "|"
-    for key, value in dict.items():
-        if isinstance(value, list):
-            value = ", ".join(value)
-        if not unrolled_items:
-            unrolled_items = f"{key}: {value}"
-        else:
-            unrolled_items = f"{unrolled_items} {separator} {key}: {value}"
-
-    return unrolled_items
-
-
-def unroll_dict_to_list(dict: dict):
-    dict_list = []
-    for key, value in dict.items():
-        if isinstance(value, list):
-            value = ", ".join(value)
-            dict_list.append(f"{key}: {value}")
-        else:
-            dict_list.append(f"{key}: {value}")
-
-    return dict_list
-
-
-def parse_json_tags(tags: list):
-    dict_tags = {}
-    if tags and tags != [{}] and tags != [None]:
-        for tag in tags:
-            if "Key" in tag and "Value" in tag:
-                dict_tags[tag["Key"]] = tag["Value"]
-            else:
-                dict_tags.update(tag)
-
-    return dict_tags
-
-
+# TODO: move this to outputs/<compliance>/models.py
 class Check_Output_MITRE_ATTACK(BaseModel):
     """
     Check_Output_MITRE_ATTACK generates a finding's output in CSV MITRE ATTACK format.
@@ -329,6 +206,7 @@ class Compliance(BaseModel):
     AssociatedStandards: List[dict]
 
 
+# TODO: move this to outputs/json_asff/models.py
 class Check_Output_JSON_ASFF(BaseModel):
     """
     Check_Output_JSON_ASFF generates a finding's output in JSON ASFF format: https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-findings-format-syntax.html
